+++ src/tests/serde.rs
@@ -41,9 +41,7 @@ fn test() {
     t!(AtomicI64, i64, I64);
     t!(AtomicU64, u64, U64);
     // TODO: serde_test doesn't support Token::{I128,U128}
-    // #[cfg(feature = "i128")]
     // t!(AtomicI128, i128, I128);
-    // #[cfg(feature = "i128")]
     // t!(AtomicU128, u128, U128);
     #[cfg(feature = "float")]
     t!(AtomicF32, f32, F32);